backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
requests==2.31.0
dnspython==2.5.0
python-dotenv==1.0.1
beautifulsoup4==4.12.3 